task submission

+++ restaurants/views.py
@@ -59,6 +59,9 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -73,8 +76,15 @@ def restaurant_create(request):
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    
+    if not (request.user == restaurant.owner or request.user.is_staff):
+       return redirect('no-access')
+    
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,13 +99,21 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+
+    if not (request.user == restaurant_obj.owner or request.user.is_staff):
+       return redirect('no-access')
+
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
         if form.is_valid():
             form.save()
             return redirect('restaurant-list')
+
     context = {
         "restaurant_obj": restaurant_obj,
         "form":form,
@@ -103,6 +121,11 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if request.user.is_staff:
+        return redirect('no-access')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def no_access(request):
+    return render(request,'no_access.html')
